pyproject.toml
[project]
name = "sauron"
version = "0.1.0"
description = "Personal Voice Intelligence System"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "watchdog>=5.0.0",
    "openai-whisper>=20240930",
    "pyannote.audio>=3.3.0",
    "torch>=2.4.0",
    "torchaudio>=2.4.0",
    "parselmouth>=0.4.4",
    "librosa>=0.10.2",
    "soundfile>=0.12.1",
    "numpy>=1.26.0",
    "scipy>=1.14.0",
    "networkx>=3.4",
    "anthropic>=0.39.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
sauron = "sauron.main:cli"
